Move check for ALTERing datatype of distribution key.
Makes more sense to perform it in the ATPrep* stage. That's where the
very similar check for changing the datatype of a partition key happens
in upstream, too.

Because we perform checks in slightly different order now, have to modify
the 'alter_table' test so that you get the same errors as in upstream, not
the errors about changing distribution key type. We have tests for changing
distribution key in 'alter_distribution_policy' test.