gpconfigurenewsegment: use forceoverwrite when retrying pg_basebackup
I really don't like how --create-slot is being handled currently in
merge branch. Trying without --create-slot and if failure retrying
with it sounds pretty ugly.

The problem it creates with tablespace is more. As the error happens
after tablespace directories are created and on failure not
cleanedup. So, next retry even if specified with --create-slot option
fails as tablespace directory is not empty. This was exposed by
gpaddmirrors behave test. I am fixing the same by adding
forceoverwrite to true on rewrite.

This should get that job green. Medium term we really need to fix this
FIXME and not rely on pg_basebackup error to detect the slot
existence.