Avoid PM_STATUS_STANDBY if promote_trigger_file is present
If promote_trigger_file is present user wishes to promote the
standby. In this case `pg_ctl -w` should wait for connections to be
allowed to database. Instead currently irrespective of
promote_trigger_file presence or not, for standby (not hotstandby) as
soon as recovery is started PM_STATUS_STANDBY is written to
postmaster.pid file. This causes the `pg_ctl -w` to stop waiting but
then connections may continue to fail as standby has not be fully
promoted.

Hence, if promote_trigger_file is present avoid writing
PM_STATUS_STANDBY. On promotion, PM_STATUS_READY will be written which
conveys the correct reality.

There is nothing GPDB specific about this. Will post patch upstream
for the same. Commiting this to GPDB now, as without this
gpactivatestandby behave tests are flaky.