Enable GDD in TAP tests.
pgbench tests with multiple clients were failing with deadlock errors.
All the other TAP tests seem to be passing without GDD at the moment,
but I think it makes sense to enable it globally. More testing of GDD
seems better, and the behavior with GDD enabled is closer to upstream,
so it seems more likely that any new upstream tests will work with GDD
than without.