Drop 'isolation_regression' database before pg_upgrade tests
I added a 'dropdb' after a successful isolation test run in an earlier
commit already. I didn't want to drop it on a failure, to avoid
destroying evidence of the failure, but let's drop it anyway before
running pg_upgrade tests because they're very noisy if the
'isolation_regression' database is present.