Remove FIXME, performance is acceptable.
I did some quick testing on my laptop, and it seems that the executor
overhead is acceptable.

To test, I did this:

    -- init
    create table parttab (a int, b int, c int, d int, e int) distributed by (b) partition by range (a) (start(1) end (4) every (1));
    insert into parttab select 1+g%3, 1, g,g,g from generate_series(1, 10000000) g;
    vacuum parttab;

    -- repeat this test many times
    \timing
    copy parttab to '/tmp/foo<SEGID>' on segment;

This is test was designed to eliminate as much of the other overhead as
possible, to make any possible executor overhead show up. I saw quite a
lot of jitter on the runtime of that test on my laptop, but it seemed to
be aroun 3.7 s on both the iteration branch and on master. I think that's
enough testing for this.