Align the handling of synchronous_standby_names GUC to upstream
This removes the Greenplum specific hack to return hardcoded value of
1 (meaning synchronous) as priority of a standby.  Greenplum mirrors
and the standby are synchronous replicas, but we can continue to use
upstream logic to determine that.

One Greenplum specific hack is still needed to indicate that the
standby should be treated as a synchronous replica of the Greenplum
master even when its name is not listed in synchronous_standby_names
GUC.  Greenplum primary and mirror pairs continue to use
synchronous_standby_names GUC like upstream, no change there.