Forbid PREPARE TRANSCATION completely.
It was already forbidden in normal mode.

It was causing the 009_twophase test to fail badly, because PREPARE
TRANSACTION succeeded, but after that recovery panicked, because the
user-supplied prepared transaction name didn't follow the expected
"%u-%u" format that GPDB uses for all distributed transactions. Now
that we forbid it altogether, the test will of course still fail, but
at least recovery will finish so that the subsequent tests are executed.