Put back fast-path versions of evaluating ScalarArrayOpExprs.
We had fast path versions of ExecEvalScalarArrayOpExpr(), for a few
built-in equality functions, but they were lost with the merge when all
the whole expression evaluation code got heavily refactored into the
opcode based approach. Performance testing shows that it still matters,
so re-implement it.

On my laptop, TPC-DS query 8 was about 2x slower on the merge branch than
on master. This fixed the regression, making it the same speed again.