Disable logging collector in TAP tests again.
In commit d72a723ccb I removed 'logging_collector=off' from the "pg_ctl
start" command line used in TAP tests. Now I know we had it:
In PostgreSQL, logging_collector is disabled by default, but in GPDB, it's
enabled. When it was enabled, the tests in src/bin/scripts/ started to
fail. The issues_sql_like() function depended on redirecting the logs,
which doesn't work with logging_collector=on.

So disable the logging_collector in TAP tests again. But this time, put
the logging_collector=off in the postgresql.conf file instead of the
pg_ctl command line, so that it can be overridden by individual tests.