Allow PREPARE TRANSACTION in utility mode
When postmaster or a backend is running in utility mode, this patch
allows a transaction identifier that does not conform to Greenplum
format ("<timestamp>-<transactionID>").  This way, we can run
009_twophase.pl tests from upstream that execute PREPARE TRANSCTION
commands.  These are useful scenarios and in utlity mode, Greenplum
should behave almost the same as upstream PostgreSQL.

PREPARE TRANSCTION command continues to be disabled when run on QD.

The following commits are reverted by this patch:

Revert "Forbid PREPARE TRANSCATION completely."
This reverts commit d2fee3a98f0f9daec19de5757a011521d991d34f.

Revert "Disable tests that used PREPARE TRANSACTION."
This reverts commit 32c2da6c808678e45d471128ed3f8c294ba8375d.