Fix dispatch logic when creating materialized views
If "WITH NO DATA" clause is specified when creating materialized view,
the query need not be executed.  CREATE TABLE AS also supports WITH NO
DATA clause and the same optimization is already implemented for it.
We seemed to be reusing the same optimization for materialized views
but that's causing CreateStmt to be dispatched to QEs in both CTAS as
well as materialized view cases.  QEs cannot distinguish between the
two and end up skipping rules population for materialize view when
"WITH NO DATA" is specified.  The rules are correctly populated on QD
because it goes through ExecCreateTableAs command.  This was caught by
gpcheckcat.

This patch disables the optimization to avoid executing query for
materialized views.  A fixme is added to remember to implement it
correctly in future.