Fix the new ALTER TABLE ADD COLUMN with "fast default".
It's important that all segments have the same value. gpcheckcat spotted
that that wasn't happening, if you did something like this:

    ALTER TABLE foo ADD COLUMN ts timestamp DEFAULT now();

Each segment evaluated now() on their own, yielding slightly different
values. That's no good, so arrange this so that QD evaluates it and
dispatches the value to the QEs.