gpcheckcat: fix the join filter for pg_attribute --> pg_class
This test joins pg_attribute with pg_class on attrelid = oid.  To
exclude tables having zero columns from this query, (relnatts > 0)
condition was added to this filter.  That condition breaks the test
when there are rows in pg_attribute but no corresponding oid in
pg_class.  This should be flagged as an error, so fix the filter
condition such that NULLs are not filtered out.