Use > rather than < in range partition pruning test
The less than (<) operator has a strategy number of 1, which is too
special because it coincides with the equal (=) strategy for hash
partitioning. In the same spirit of the test, switch to using greater
than (>) which has strategy number 5.

While we're at it, also fix a typo for testing that we are falling back
for list partitioning.