Force explain in some regress tests
Plan cost changes are usually not considered as diffs because atmsort.pl
can help to ignore them, however if the "explain" keyword is put in a
separate line with the query, then atmsort.pl does not treat the block
as a plan and plan cost changes will cause the test to fail.

To make atmsort.pl work, there are two choices, one is to move the
"explain" keyword to the same line of the query, or we could simply put
a "-- force_explain" line before the query, so atmsort.pl treats the
block as a plan forcefully.