Remove FIXMEs that are no longer relevant
A lot of code churn has taken place on the iteration branch since
these fixmes were added.  The regression tests create_table_like,
ao_locks are now passing without the additional AccessShareLock
workaround.

The failure of isolation2 tests uao/parallel_delete and
uao/parallel_update brought these fixmes under the scanner.  With Orca
turned on, the tests used to fail because the AccessShareLock, as
expected in the answer file, was not found to be held when a DELETE
statement was executed on AO table.  The lock was found to be relased
by the relation_close() call fixed by this patch.