Update uao/parallel_{delete|update} isolation2 test anser files
Orca specific answer files for the tests are updated to exclude
AccessShareLock from the list of locks expected to be held on a AO
table undergoing DELETE or UPDATE operation.

As it stands on master branch, the AccessShareLock is acquired on QD
when Orca is enabled because of the following reasons:

(1) For a statement such as "DELETE FROM ao WHERE a = 10", Orca
generates a plan with two range table entries, one representing the
table being modified and the other for the table being scanned.  In
the example query, both RTEs represent the same relation.

(2) ExclusiveLock is held during planning for the relation being
modified.  But for the RTE entry corresponding to the relation to be
scanned, AccessShareLock is explicitly acquired in
ExecOpenScanRelation().

On the iteration branch, the Orca behavior is unchanged - we still get
a plan with two RTEs.  However, ExecOpenScanRelation() is changed such
that the lock mode is hardcoded to NoLock when opening the relation to
be scanned.  Therefore, Orca specific answer files need to be updated
accordingly.