Remove Orca answer file as locks acquired no longer differ
The test executes DELETE statement against the root partitioned table
and checks the number of locks held on QD.  Previously, the number of
locks held with Orca enabled was one less than without Orca.  This is
no longer the case.  Locks acquired on the root table are
RowExclsiveLock during parsing and AccessShareLock during execution,
specifically in InitResultRelInfo, when partition constraint is
obtained.  This logic is newly introduced in iteration branch, as part
of table partitioning feature.  The implementation of table
partitioning in master branch is entirely Greenplum specific and does
not involve the addtional AccessShareLock lock with Orca enabled.  I
didn't investigate why the number of locks held differs with Orca
enabled on master branch but on iteration branch, the results are
identical and they look correct.  So let's remove the Orca specific
answer file.