Fix flaky brin_ao regress test
In the brin_ao regress test we create a table brintest_ao with a brin
index, then performs brin indexscan on different columns to verify the
correctness.  After that we also perform some vacuum tests, first vacuum
that table, then copy all the data to itself, and finally vacuum again.

There is also a similar brin_aocs regress test, the only difference is
that its table is brintest_aocs.

However there is a typo in brin_aocs, it vacuums brintest_ao instead of
brintest_aocs, as well as the data doubling.  So in brin_ao we can
randomly see below message:

    +WARNING:  unexpected number of results 106 for (int4rangecol,>,int4range,empty,53)

The actual number of results is double size of the expected one, that's
simply because that the data does is doubled concurrently.

Corrected the typo, both the brin_ao and brin_aocs should be stable now.

Co-authored-by: Pengzhou Tang <ptang@pivotal.io>
Co-authored-by: Ning Yu <nyu@pivotal.io>