Fix regress::join_hash
It adjusts work_mem to trigger multi-batch hashjoin and batch expansion.
In gpdb we need to adjust statement_mem instead of work_mem, however the
minimal possible value of statement_mem differs depends on cassert is
enable or not, with cassert it is 50kB, without cassert it is 1000kB.

We used to set statement_mem to 512kB in the test, so it only work with
a cassert enabled build.  We have to adjust it to 1000kB, and the table
sizes are also adjusted to meet the test purpose.