Fix isolation2::gdd/dml_locks_only_targeted_table_in_query
The macro RELCACHE_FORCE_RELEASE is defined iff USE_ASSERT_CHECKING is
defined, and when RELCACHE_FORCE_RELEASE is defined the relcache is
forcefully released when closing the relation.

The function generate_partition_qual() will behave differently depends
on the existence of the relcache.

- if the relation is not cached, it will open it in AccessShareLock
  mode, and save the relpartbound in the relcache;
- if the relation is already cached, it will load the relpartbound from
  the cache directly without opening the relation;

So as a result, in the following transactions we will see an extra
AccessShareLock lock in a --enable-cassert build compared to a
--disable-cassert build.

To make the test results stable, we do not reuse the sessions in the
test, all the tests are performed without the relcache.