Disable ORCA in gangsize test
Running the gangsize test with ORCA in CI produces cache lookup
failures. We suspect the cause of this is in the collect_tabstat and
this isn't an issue with gangsize, but this test does expose a
legitimate issue that needs to be fixed. For now, we disable ORCA for
this test until this can be debugged and add a GPDB_12_MERGE_FIXME.