Fix gporca test for ORCA
A few queries set the `optimizer_plan_id` to a specific value. Since we
were falling back, whenever the space size changed (which can be quite
common as ORCA may explore more or fewer alternatives without any impact
on the plan chosen), the output of this test would also change.
However, even the optimizer_plan_id is not necessarily stable and the
intended plan isn't evident anymore. Remove this setting, and in the
future, we should force plans in tests by setting GUCs rather than
forcing a plan ID.