Disable ORCA for bfv_partition tests
ORCA currently produces wrong results/possibility of crashing if the
constant used in the equal function is a different type than the
partition boundry. For now, we're disabling the test until after the
merge, but this will be picked up right after. Falling back may take a
similar amount of effort as fixing the issue.