Revert a3dd474470 and let log_line_prefix GUC be defunct
It was a mistage to change the defunct status of log_line_prefix
because we don't allow changing it in Greenplum.  Commit a3dd474470
also changed defunct status of "autovacuum" GUC but its defunct status
was legitimately removed on master branch by the autostats commit
0b17968e80.  The "autovacuum" GUC is therefore left unchanged in this
revert.

Thanks Heikki for spotting this mistake in the ocean of commits on the
iteration 12 branch.