Avoid calling RelationInitTableAccessMethod() for partition roots (#62)
The call was originally intended to ensure the AM can be passed down to
partition leaves from the root. RelationInitTableAccessMethod() ensured
that the table AM handler was set and decisions based on that were taken in other
pieces of code.

However, we no longer rely on the table AM handler, but the relam property
instead, which is not touched by this routine. Thus, this call is superfluous.

It also breaks the following test case in update.sql:
@@ -894,7 +891,11 @@ insert into utrtest values (2, 'bar')
   returning *, tableoid::regclass, cmin;  -- fails
-ERROR:  cannot retrieve a system column in this context
+ a |  b  | tableoid | cmin
+---+-----+----------+------
+ 2 | bar | utr2     |    0
+(1 row)
+

So, remove it.

Co-authored-by: Soumyadeep Chakraborty <soumyadeep2007@gmail.com>